refactor(web/stores): extract useCompactStore from useAppStore
Move every compact-mode field — compactMode, compactAlwaysOnTop,
compactSize, compactFontSize, compactAmbientIntensity, compactShow*,
compactDefaultAlwaysOnTop — together with their setters, the
toggle/reset actions, the IPC-touching setCompactMode rollback dance,
the post-rehydrate Electron sync, and the COMPACT_DIMENSIONS / CMP_*
constants into a dedicated useCompactStore.

The new store persists to its own bucket (shiranami.compact-store) and
runs a one-shot import-time migration that lifts the compact fields
out of the pre-split shiranami.app-store key on first load. The legacy
per-key importLegacyAppStore() in useAppStore.ts still funnels the
shiranami.compact-always-on-top key through the combined bucket so
that path's behavior is preserved end-to-end (the per-key migration
writes into shiranami.app-store, then useCompactStore reads it from
there).

Direct compact consumers (App.tsx, CompactPlayer, CompactSection,
CompactModePreview, PlayerBar, PlayerOverflowMenu) now read from
useCompactStore. The Ctrl+Shift+M / Ctrl+Shift+T keyboard shortcuts
route through useCompactStore. Compact tests move to a dedicated
useCompactStore.test.ts (12 behavior cases plus 3 legacy-import
cases). useAppStore re-exports compact constants/types so the few
remaining indirect imports keep compiling through the next commit.

Note: COMPACT_DIMENSIONS still lives next to compact state (now in
useCompactStore.ts) per the comment requesting the constants stay
matched with apps/desktop/src/main/ipc/window.ts. The contracts-
package migration is intentionally out of scope here.
